added unit tests
I have added some tests with the default value of 0.
I hope this fixes #111 and fixes #24

+++ src/NLog/Targets/FileTarget.cs
@@ -869,7 +869,7 @@ namespace NLog.Targets
 
         private void RecursiveRollingRename(string fileName, string pattern, int archiveNumber)
         {
-            if (archiveNumber >= this.MaxArchiveFiles)
+            if (this.MaxArchiveFiles != 0 && archiveNumber >= this.MaxArchiveFiles)
             {
                 File.Delete(fileName);
                 return;
@@ -955,7 +955,7 @@ namespace NLog.Targets
                 nextNumber = 0;
             }
 
-            if (minNumber != -1)
+            if (minNumber != -1 && this.MaxArchiveFiles != 0)
             {
                 int minNumberToKeep = nextNumber - this.MaxArchiveFiles + 1;
                 for (int i = minNumber; i < minNumberToKeep; ++i)
@@ -1023,12 +1023,15 @@ namespace NLog.Targets
                     }
                 }
 
-                for (int fileIndex = 0; fileIndex < filesByDate.Count; fileIndex++)
+                if (this.MaxArchiveFiles != 0)
                 {
-                    if (fileIndex > files.Count - this.MaxArchiveFiles)
-                        break;
+                    for (int fileIndex = 0; fileIndex < filesByDate.Count; fileIndex++)
+                    {
+                        if (fileIndex > files.Count - this.MaxArchiveFiles)
+                            break;
 
-                    File.Delete(filesByDate[fileIndex]);
+                        File.Delete(filesByDate[fileIndex]);
+                    }
                 }
             }
             catch (DirectoryNotFoundException)
